Final Version of Memory Card Game commented
The animation to flip cards is too complicated at the moment, so I decided to upload the final version of this project but I don't give up. I think it'll be useful to comment pretty much each line of code (especialy javascript) to show that I understand what is going on.
So I will find these commented files to help me in other similar project in the future.